fix(runtime): fix config error db lost connect
